+++ sdk/include/spatial/rendering/DebugVertex.h
@@ -1,0 +1,13 @@
+#pragma once
+
+#include "spatial/core/Vec3.h"
+#include "spatial/rendering/Color.h"
+
+namespace spatial::rendering
+{
+    struct DebugVertex
+    {
+        core::Vec3 position;
+        Color color;
+    };
+}